fix bug in mergeBases.
This bug was probably responsible for the weird conflicts that have
shown up from time to time.  We were satisfied with a merge base that
is ancestor to all of the heads, but isn't ancestor to some of the
ancestors of those heads, which meant that some parts of the history
were being clumped together.

The new mergeBase only accepts bases that are "choke points" where all
patches are either their descendent or their ancestor.

The algorithm used is terribly wasteful, but such is life...

Tested-on: collins x86_64 GNU/Linux